refactor!: move Breadcrumb into the Core namespace
PageSchema lives in Core and now sets breadcrumb overrides directly,
so Breadcrumb can no longer live in Http without Core depending upward
on the orchestration layer, which the layering rules forbid.

BREAKING CHANGE: Page::breadcrumbs() now returns array<int, Breadcrumb>
instead of the array{title: string, href: string} shape, and Breadcrumb
moved from Lattice\Lattice\Http to Lattice\Lattice\Core. Existing array-shaped
breadcrumbs still serialize identically on the wire, so the failure is
static-analysis-only — update the return type and the import.
