+++ src/Core/PageSchema.php
@@ -3,7 +3,6 @@ declare(strict_types=1);
 
 namespace Lattice\Lattice\Core;
 
-use Lattice\Lattice\Http\Breadcrumb;
 use Lattice\Lattice\Ui\Components\Component;
 use Lattice\Lattice\Ui\Concerns\FiltersRenderableComponents;
 use Lattice\Lattice\Ui\Concerns\ResolvesSchemaEntries;
