+++ tests/Feature/Core/BreadcrumbTest.php
@@ -3,8 +3,8 @@ declare(strict_types=1);
 
 use Illuminate\Support\Facades\Route;
 use Lattice\Lattice\Attributes\AsPage;
+use Lattice\Lattice\Core\Breadcrumb;
 use Lattice\Lattice\Core\PageSchema;
-use Lattice\Lattice\Http\Breadcrumb;
 use Lattice\Lattice\Http\Page;
 use Lattice\Lattice\Ui\Components\Text;
